# docs/design/seat-reservation/design-api-contract.yaml
openapi: 3.0.3
info:
  title: Seat Reservation API
  version: 1.0.0
  description: |
    Public seat reservation platform API.
    All state-mutating endpoints require authentication via HttpOnly JWT cookie
    set by POST /api/auth/login.
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Local development

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
# Authentication is carried in an HttpOnly cookie named `token`.
# The cookie is set by POST /api/auth/login and cleared by POST /api/auth/logout.
# No Authorization header is used — the browser sends the cookie automatically.
# Routes marked with cookieAuth require a valid, non-expired JWT.
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
...
    ReservationStatus:
      type: string
      enum: [PENDING_PAYMENT, CONFIRMED, EXPIRED]
    Reservation:
      type: object
      properties:
        id:
          type: integer
          example: 42
        userId:
          type: integer
          example: 1
        seatId:
          type: integer
          example: 2
        status:
          $ref: '#/components/schemas/ReservationStatus'
        paymentAttemptCount:
          type: integer
          example: 0
        expiresAt:
          type: string
          format: date-time
          example: '2026-06-22T10:10:00Z'
        createdAt:
          type: string
          format: date-time
          example: '2026-06-22T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-06-22T10:00:00Z'
    CreateReservationRequest:
      type: object
      required: [seatId]
      properties:
        seatId:
          type: integer
          example: 2

    # --- Payment ------------------------------------------------------------
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Seat is not available

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /reservations:
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Reserve a seat
      description: |
        Creates a reservation for the given seat.

        **Race condition safety:** The seat row is locked with `SELECT FOR UPDATE`
        inside a transaction. If two concurrent requests target the same seat,
        only the first succeeds; the second receives 409.

        **Preconditions checked:**
        - Seat must be AVAILABLE.
        - The authenticated user must not already have a PENDING_PAYMENT reservation.

        On success, the reservation has status `PENDING_PAYMENT` and `expires_at`
        set to 10 minutes from now.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
      responses:
        '201':
          description: Reservation created
          content:
            application/json:
              schema:
                type: object
                properties:
                  reservation:
                    $ref: '#/components/schemas/Reservation'
        '400':
          description: User already has a pending reservation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: You already have a pending reservation
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Unauthorized
        '409':
          description: Seat is not available (taken or race condition)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Seat is not available
...